requirements.txt
django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
aiogram==3.0.0b7
aiohttp==3.8.4